+++ rust/crd/src/lib.rs
@@ -1,9 +1,11 @@
 pub mod affinity;
 pub mod authentication;
+mod git_sync;
 
 use crate::affinity::get_affinity;
 use crate::authentication::AirflowAuthentication;
 
+use git_sync::GitSync;
 use serde::{Deserialize, Serialize};
 use snafu::{OptionExt, ResultExt, Snafu};
 use stackable_operator::{
@@ -47,6 +49,7 @@ pub const GIT_CONTENT: &str = "content-from-git";
 pub const GIT_ROOT: &str = "/tmp/git";
 pub const GIT_LINK: &str = "current";
 pub const GIT_SYNC_NAME: &str = "gitsync";
+pub const GIT_SAFE_DIR: &str = "safe.directory";
 
 pub const TEMPLATE_VOLUME_NAME: &str = "airflow-executor-pod-template";
 pub const TEMPLATE_CONFIGMAP_NAME: &str = "airflow-executor-pod-template";
@@ -209,52 +212,6 @@ impl CurrentlySupportedListenerClasses {
             CurrentlySupportedListenerClasses::ExternalUnstable => "NodePort".to_string(),
             CurrentlySupportedListenerClasses::ExternalStable => "LoadBalancer".to_string(),
         }
-    }
-}
-
-#[derive(Clone, Debug, Default, Deserialize, JsonSchema, PartialEq, Eq, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct GitSync {
-    pub repo: String,
-    pub branch: Option<String>,
-    pub git_folder: Option<String>,
-    pub depth: Option<u8>,
-    pub wait: Option<u16>,
-    pub credentials_secret: Option<String>,
-    pub git_sync_conf: Option<BTreeMap<String, String>>,
-}
-
-impl GitSync {
-    pub fn get_args(&self) -> Vec<String> {
-        let mut args: Vec<String> = vec![];
-        args.extend(vec![
-            "/stackable/git-sync".to_string(),
-            format!("--repo={}", self.repo.clone()),
-            format!(
-                "--branch={}",
-                self.branch.clone().unwrap_or_else(|| "main".to_string())
-            ),
-            format!("--depth={}", self.depth.unwrap_or(GIT_SYNC_DEPTH)),
-            format!("--wait={}", self.wait.unwrap_or(GIT_SYNC_WAIT)),
-            format!("--dest={GIT_LINK}"),
-            format!("--root={GIT_ROOT}"),
-            format!("--git-config=safe.directory:{GIT_ROOT}"),
-        ]);
-        if let Some(git_sync_conf) = self.git_sync_conf.as_ref() {
-            for (key, value) in git_sync_conf {
-                // config options that are internal details have
-                // constant values and will be ignored here
-                if key.eq_ignore_ascii_case("--dest")
-                    || key.eq_ignore_ascii_case("--root")
-                    || key.eq_ignore_ascii_case("--git-config")
-                {
-                    tracing::warn!("Config option {:?} will be ignored...", key);
-                } else {
-                    args.push(format!("{key}={value}"));
-                }
-            }
-        }
-        args
     }
 }
 
@@ -785,99 +742,5 @@ mod tests {
         assert_eq!("KubernetesExecutor", cluster.spec.executor.to_string());
         assert!(cluster.spec.cluster_config.load_examples.unwrap_or(false));
         assert!(cluster.spec.cluster_config.expose_config.unwrap_or(false));
-    }
-
-    #[test]
-    fn test_git_sync() {
-        let cluster = "
-        apiVersion: airflow.stackable.tech/v1alpha1
-        kind: AirflowCluster
-        metadata:
-          name: airflow
-        spec:
-          image:
-            productVersion: 2.7.2
-          clusterConfig:
-            loadExamples: false
-            exposeConfig: false
-            credentialsSecret: simple-airflow-credentials
-            dagsGitSync:
-              - name: git-sync
-                repo: https://github.com/stackabletech/airflow-operator
-                branch: feat/git-sync
-                wait: 20
-                gitSyncConf: {}
-                gitFolder: tests/templates/kuttl/mount-dags-gitsync/dags
-          webservers:
-            roleGroups:
-              default:
-                config: {}
-          celeryExecutors:
-            roleGroups:
-              default:
-                config: {}
-          schedulers:
-            roleGroups:
-              default:
-                config: {}
-          ";
-
-        let deserializer = serde_yaml::Deserializer::from_str(cluster);
-        let cluster: AirflowCluster =
-            serde_yaml::with::singleton_map_recursive::deserialize(deserializer).unwrap();
-
-        assert!(cluster.git_sync().is_some(), "git_sync was not Some!");
-        assert_eq!(
-            Some("tests/templates/kuttl/mount-dags-gitsync/dags".to_string()),
-            cluster.git_sync().unwrap().git_folder
-        );
-    }
-
-    #[test]
-    fn test_git_sync_config() {
-        let cluster = "
-        apiVersion: airflow.stackable.tech/v1alpha1
-        kind: AirflowCluster
-        metadata:
-          name: airflow
-        spec:
-          image:
-            productVersion: 2.7.2
-          clusterConfig:
-            loadExamples: false
-            exposeConfig: false
-            credentialsSecret: simple-airflow-credentials
-            dagsGitSync:
-              - name: git-sync
-                repo: https://github.com/stackabletech/airflow-operator
-                branch: feat/git-sync
-                wait: 20
-                gitSyncConf:
-                  --rev: c63921857618a8c392ad757dda13090fff3d879a
-                gitFolder: tests/templates/kuttl/mount-dags-gitsync/dags
-          webservers:
-            roleGroups:
-              default:
-                config: {}
-          celeryExecutors:
-            roleGroups:
-              default:
-                config: {}
-          schedulers:
-            roleGroups:
-              default:
-                config: {}
-          ";
-
-        let deserializer = serde_yaml::Deserializer::from_str(cluster);
-        let cluster: AirflowCluster =
-            serde_yaml::with::singleton_map_recursive::deserialize(deserializer).unwrap();
-
-        assert!(cluster
-            .git_sync()
-            .unwrap()
-            .get_args()
-            .iter()
-            .any(|c| c == "--rev=c63921857618a8c392ad757dda13090fff3d879a"));
     }
 }